+++ src/TreeHouse/IoBundle/EventListener/ImportLoggingSubscriber.php
@@ -8,6 +8,7 @@ use TreeHouse\Feeder\Event\FailedItemModificationEvent;
 use TreeHouse\Feeder\Event\ResourceEvent;
 use TreeHouse\Feeder\Event\TransportEvent;
 use TreeHouse\Feeder\FeedEvents;
+use TreeHouse\IoBundle\Import\Event\ExceptionEvent;
 use TreeHouse\IoBundle\Import\Event\FailedItemEvent;
 use TreeHouse\IoBundle\Import\Event\ImporterEvent;
 use TreeHouse\IoBundle\Import\Event\PartEvent;
@@ -46,6 +47,7 @@ class ImportLoggingSubscriber implements EventSubscriberInterface
             ImportEvents::ITEM_FAILED => 'onItemFailed',
             ImportEvents::ITEM_SKIPPED => 'onItemSkipped',
             ImportEvents::PART_CREATED => 'onPartCreated',
+            ImportEvents::EXCEPTION => 'onException',
         ];
     }
 
@@ -161,17 +163,30 @@ class ImportLoggingSubscriber implements EventSubscriberInterface
         $importer = $event->getImporter();
         $result = $importer->getResult();
 
+        $total = $result->getTotal();
+        $processed = $result->getProcessed();
+        $percentage = $total > 0 ? ($processed / $total * 100) : 0;
+
         $this->logger->info(sprintf('Import ended in %s seconds', round($result->getElapsedTime())));
         $this->logger->info(
             sprintf(
                 'Processed <info>%s</info> of <info>%s</info> items (<info>%d%%</info>):',
-                $result->getProcessed(),
-                $result->getTotal(),
-                $result->getProcessed() / $result->getTotal() * 100
+                $processed,
+                $total,
+                $percentage
             )
         );
+
         $this->logger->info(sprintf('- succes:  <info>%s</info>', $result->getSuccess()));
         $this->logger->info(sprintf('- failed:  <info>%s</info>', $result->getFailed()));
         $this->logger->info(sprintf('- skipped: <info>%s</info>', $result->getSkipped()));
+    }
+
+    /**
+     * @param ExceptionEvent $event
+     */
+    public function onException(ExceptionEvent $event)
+    {
+        $this->logger->error($event->getException()->getMessage());
     }
 }
